+++ src/components/common/feedItem/feedModal/style.ts
